requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==4.3.0
pydantic==2.10.1
pytest==8.3.3
python-dotenv==1.0.1
selene==2.0.0rc9
selenium==4.26.1
